Create modules HR and IOD
[ver 0_9_4_1]

HR module

- Created dictionaries: Locations, Professional groups, Organizational
units, Positions in the module HR
- Added functionality Employee - Basic data
- Added functionality Employee - Employment

IOD module

- Created a register of processing activities in the IOD module

Plans

- Changed functionality downloading of the coordinator plans in the
Coordinator, Public Procurement, Accountant and Director modules.
Currently, plans are downloading by the current year. Changing the
year in the selection field automatically downloads the plan for the
given year.
- Added a plan type column when exporting a coordinator plan to XLSX in
the Accounting module
- Changed the display of financial and investment plan items when
recording invoice items. Currently, the availability of items from the
plans is download based on the year from the date of sale in the invoice
header, and not based on the current year as before.This enables the
recording of invoices and the implementation of financial and investment
plans regardless of the current year, provided that there is an approved
or realized plan of a given type of coordinator.
- Fixed an error when saving a supposition in the coordinator's plan at
the editing mode.
- Added approval of the public procurement plan of an institution at the
moment of approval by the Chief of the last plan of the coordinator
included in the plan of the institution
- Fixed updating of the position status during the process of updating
the financial plan in the Accountant module in the item correction mode
if the value of the accepted correction was equal to the requested
value. Current status is SK, previously ZA. This resulted in errors in
the value of plan items at the time of creating / modifying the
institution's financial plan.
- Fixed an error when adding a new item to the financial plan update and
public procurement.The error caused that when adding more than one item
to the plan, a new item was not created, but the last added item of the
plan was modified.
- Added the functionality of previewing invoices that make up the amount
of implementation of a given item in the financial plan of the
institution
- Added the option of exporting invoices that make up the realization
amount of a given item in the institution's financial plan to an XLS
file

Public procurement applications

- Increasing the number of characters to 2000 for the purchase
justification field on the public procurement application.
- Fixed functionality of closing applications in the public procurement
module. Depending on whether the application contains parts or not, the
value of submitted applications within a given assortment group in the
plan of the institution and the plan of the coordinator is released at
the moment marking the part as unrealized. If the application does not
contain a part, the value of submitted applications is released at the
moment of completion or complete cancellation of the application. In
addition, dependent items in the plan of the organization and the plan
of the coordinator are also searched and the values are also updated.
- Added filtering of public procurement applications based on the date
of transfer to the Public Procurement in the Public Procurement,
Accountant and Director modules

Public procurement applications protocol

- Fixed the permission to access the list of protocols in the Public
Procurement module
- Fixed the permission to access the list of protocols in the Director
module
- Fixed inactive Price field of the selected offer in the Public
Procurement protocol
- Fixed protocol details retrieving on close and reopening details
protocol from the protocols menu in the Public Procurement Accountant
and Director modules.
- Fixed of "white page" error related to displaying protocol details
from the level of the application from a module other than the
Coordinator, in a situation where the application has been sent by the
Coordinator and the protocol for the application has not yet been
created.

Contracts

- Fixed the filtering of the list of contracts with the use of the
fields number and subject of the contract, the current search values
remain visible in the given field.
- Added validation of exceeding the contract value in a situation where
there was a modification of the amount of implementation in previous
years
- Added validation for exceeding the value of the contract when the
invoice was added using the invoice functionality and then the correct
contract is added in the edit mode.

Invoices

- Fixed the functionality of adding an invoice to a contract when is
added when adding a new contract
- Fixed of the validation of exceeding the realization of the contract
in the situation when an invoice item was added using the entire
available value of the contract. Before the correction, it was possible
to record an amount lower by 1 than the total available amount of the
contract
- Fixed validation of the required field Name when registering invoice
items
- Fixed displaying of an error message for the net amount field of an
invoice item when modifying the gross amount in the event of exceeding
the net amount the coordinator's plan position
- Fixed search invoices in the situation for search all invoices without
division by year

Permissions

- Added permission to create a dictionary item in the Public Procurement
module
- Added permission to removal a dictionary item in the Public
Procurement module

Components

- Modified sorting of text type columns in the tableForm and guiForm
components has been, currently the sorting is not case-sensitive
- Fixed display of the dialog when deleting items for the dictionary
component
- Added validation if there is an item with the given name in the
dictionary

+++ src/main/java/pl/viola/ems/service/modules/publicProcurement/plans/institution/impl/PublicInstitutionPlanServiceImpl.java
@@ -29,7 +29,6 @@ import pl.viola.ems.payload.modules.publicProcurement.plans.PublicInstitutionPla
 import pl.viola.ems.payload.modules.publicProcurement.plans.PublicInstitutionPlanPositionResponse;
 import pl.viola.ems.payload.modules.publicProcurement.plans.PublicInstitutionPlanResponse;
 import pl.viola.ems.service.common.JasperPrintService;
-import pl.viola.ems.service.modules.coordinator.plans.PlanService;
 import pl.viola.ems.service.modules.publicProcurement.plans.institution.PublicInstitutionPlanService;
 import pl.viola.ems.utils.Utils;
 
@@ -61,15 +60,16 @@ public class PublicInstitutionPlanServiceImpl implements PublicInstitutionPlanSe
     CoordinatorPlanRepository coordinatorPlanRepository;
 
     @Autowired
-    PlanService coordinatorPlanServices;
-
-    @Autowired
     InstitutionCoordinatorPlanPositionRepository institutionCoordinatorPlanPositionRepository;
 
     @Override
-    public List<PublicInstitutionPlanResponse> getPlans(String levelAccess) {
-        List<PublicInstitutionPlanResponse> publicInstitutionPlans = new ArrayList<>();
-        List<InstitutionPlan> institutionPlans = institutionPlanRepository.findByType(CoordinatorPlan.PlanType.PZP);
+    public Set<PublicInstitutionPlanResponse> getPlans(final int year, String levelAccess) {
+
+        Set<PublicInstitutionPlanResponse> publicInstitutionPlans = new HashSet<>();
+
+        Set<InstitutionPlan> institutionPlans =
+                year != 0 ? institutionPlanRepository.findByYearAndTypeInOrderByIdDesc(year, Collections.singletonList(CoordinatorPlan.PlanType.PZP))
+                        : institutionPlanRepository.findByType(CoordinatorPlan.PlanType.PZP);
 
         if (!institutionPlans.isEmpty()) {
             institutionPlans.forEach(institutionPlan -> publicInstitutionPlans.add(setPlanAmountValues(institutionPlan)));
@@ -209,12 +209,11 @@ public class PublicInstitutionPlanServiceImpl implements PublicInstitutionPlanSe
         } else {
             List<CoordinatorPlan> coordinatorPlans = coordinatorPlanRepository.findByStatusAndTypeAndYear(CoordinatorPlan.PlanStatus.AN, plan.getType(), plan.getYear());
 
-            coordinatorPlans.forEach(coordinatorPlan -> {
-                coordinatorPlan.setStatus(CoordinatorPlan.PlanStatus.ZA);
-            });
+            coordinatorPlans.forEach(coordinatorPlan -> coordinatorPlan.setStatus(CoordinatorPlan.PlanStatus.ZA));
 
             coordinatorPlanRepository.saveAll(coordinatorPlans);
-            plan.setStatus(InstitutionPlan.InstitutionPlanStatus.AD);
+            plan.setApproveUser(action.equals("approve") ? user : null);
+            plan.setStatus(InstitutionPlan.InstitutionPlanStatus.ZA);
         }
         return institutionPlanRepository.save(plan).getStatus();
     }
@@ -222,7 +221,9 @@ public class PublicInstitutionPlanServiceImpl implements PublicInstitutionPlanSe
     @Override
     @Transactional
     public void updateInstitutionPublicProcurementPlan(CoordinatorPlan coordinatorPlan) {
-        InstitutionPlan institutionPlan = institutionPlanRepository.findByYearAndTypeAndStatusIn(coordinatorPlan.getYear(), CoordinatorPlan.PlanType.PZP, Collections.singletonList(InstitutionPlan.InstitutionPlanStatus.UT));
+        InstitutionPlan institutionPlan = institutionPlanRepository.findByYearAndTypeAndStatusIn(
+                coordinatorPlan.getYear(), CoordinatorPlan.PlanType.PZP, Collections.singletonList(InstitutionPlan.InstitutionPlanStatus.UT)
+        ).orElse(null);
 
         //Institution plan not exist
         if (institutionPlan == null) {
@@ -230,7 +231,9 @@ public class PublicInstitutionPlanServiceImpl implements PublicInstitutionPlanSe
                 Create copy Institution plan on approve chief Public Procurement first coordinator plan update
             */
 
-            InstitutionPlan oldInstitutionPlan = institutionPlanRepository.findByYearAndTypeAndStatusIn(coordinatorPlan.getYear(), CoordinatorPlan.PlanType.PZP, Collections.singletonList(InstitutionPlan.InstitutionPlanStatus.AD));
+            InstitutionPlan oldInstitutionPlan = institutionPlanRepository.findByYearAndTypeAndStatusIn(
+                    coordinatorPlan.getYear(), CoordinatorPlan.PlanType.PZP, Arrays.asList(InstitutionPlan.InstitutionPlanStatus.ZA, InstitutionPlan.InstitutionPlanStatus.RE)
+            ).orElseThrow(() -> new AppException("Public.institution.plan.invalidStatus", HttpStatus.NOT_FOUND));
             oldInstitutionPlan.setStatus(InstitutionPlan.InstitutionPlanStatus.AA);
             oldInstitutionPlan.getPlanPositions().forEach(planPositions -> planPositions.setStatus(CoordinatorPlanPosition.PlanPositionStatus.AA));
             institutionPlanRepository.save(oldInstitutionPlan);
@@ -331,16 +334,25 @@ public class PublicInstitutionPlanServiceImpl implements PublicInstitutionPlanSe
         coordinatorPlan.getPositions().forEach(coordinatorPlanPosition -> {
             if (institutionPlanUpd.getPlanPositions().stream().anyMatch(institutionPlanUpdPosition -> institutionPlanUpdPosition.getAssortmentGroup().equals(coordinatorPlanPosition.getAssortmentGroup()))) {
                 InstitutionPlanPosition institutionPlanPosition = institutionPlanUpd.getPlanPositions().stream().filter(institutionPlanUpdPosition -> institutionPlanUpdPosition.getAssortmentGroup().equals(coordinatorPlanPosition.getAssortmentGroup())).findFirst().orElse(null);
-                InstitutionCoordinatorPlanPosition oldInstitutionCoordinatorPlanPosition = institutionPlanPosition.getInstitutionCoordinatorPlanPositions().stream().filter(institutionCoordinatorPlanPosition -> institutionCoordinatorPlanPosition.getCoordinatorPlanPosition().getId().equals(coordinatorPlanPosition.getCorrectionPlanPosition().getId())).findFirst().orElse(null);
-                InstitutionCoordinatorPlanPosition correctedInstitutionCoordinatorPlanPosition = institutionPlanPosition.getInstitutionCoordinatorPlanPositions().stream().filter(institutionCoordinatorPlanPosition -> institutionCoordinatorPlanPosition.getCoordinatorPlanPosition().getId().equals(coordinatorPlanPosition.getCorrectionPlanPosition().getId()) && coordinatorPlanPosition.getAmountRequestedNet().compareTo(coordinatorPlanPosition.getCorrectionPlanPosition().getAmountRequestedNet()) != 0).findFirst().orElse(null);
-                institutionPlanPosition.getInstitutionCoordinatorPlanPositions().add(new InstitutionCoordinatorPlanPosition(coordinatorPlanPosition, institutionPlanPosition));
-                if (correctedInstitutionCoordinatorPlanPosition != null) {
-                    institutionPlanPosition.setStatus(CoordinatorPlanPosition.PlanPositionStatus.WY);
+                if (institutionPlanPosition != null) {
+                    if (coordinatorPlanPosition.getCorrectionPlanPosition() != null) {
+                        /*Sytuacja gdy istnieje grupa w planie instytucji u innego koordynatora o obecy dopiero dodaje grupę podczas aktualizacji */
+                        InstitutionCoordinatorPlanPosition oldInstitutionCoordinatorPlanPosition = institutionPlanPosition.getInstitutionCoordinatorPlanPositions().stream().filter(institutionCoordinatorPlanPosition -> institutionCoordinatorPlanPosition.getCoordinatorPlanPosition().getId().equals(coordinatorPlanPosition.getCorrectionPlanPosition().getId())).findFirst().orElse(null);
+                        InstitutionCoordinatorPlanPosition correctedInstitutionCoordinatorPlanPosition = institutionPlanPosition.getInstitutionCoordinatorPlanPositions().stream().filter(institutionCoordinatorPlanPosition -> institutionCoordinatorPlanPosition.getCoordinatorPlanPosition().getId().equals(coordinatorPlanPosition.getCorrectionPlanPosition().getId()) && coordinatorPlanPosition.getAmountRequestedNet().compareTo(coordinatorPlanPosition.getCorrectionPlanPosition().getAmountRequestedNet()) != 0).findFirst().orElse(null);
+                        if (correctedInstitutionCoordinatorPlanPosition != null) {
+                            institutionPlanPosition.setStatus(CoordinatorPlanPosition.PlanPositionStatus.WY);
+                        }
+                        institutionPlanPosition.getInstitutionCoordinatorPlanPositions().remove(oldInstitutionCoordinatorPlanPosition);
+                        if (oldInstitutionCoordinatorPlanPosition != null) {
+                            institutionCoordinatorPlanPositionRepository.delete(oldInstitutionCoordinatorPlanPosition);
+                        }
+                    } else {
+                        institutionPlanPosition.setStatus(CoordinatorPlanPosition.PlanPositionStatus.WY);
+                    }
+                    institutionPlanPosition.getInstitutionCoordinatorPlanPositions().add(new InstitutionCoordinatorPlanPosition(coordinatorPlanPosition, institutionPlanPosition));
+                    institutionPlanPosition.setAmountRequestedNet(institutionPlanPosition.getInstitutionCoordinatorPlanPositions().stream().map(InstitutionCoordinatorPlanPosition::getAmountRequestedNet).reduce(BigDecimal.ZERO, BigDecimal::add));
+                    institutionPlanPosition.setAmountRequestedGross(institutionPlanPosition.getInstitutionCoordinatorPlanPositions().stream().map(InstitutionCoordinatorPlanPosition::getAmountRequestedGross).reduce(BigDecimal.ZERO, BigDecimal::add));
                 }
-                institutionPlanPosition.getInstitutionCoordinatorPlanPositions().remove(oldInstitutionCoordinatorPlanPosition);
-                institutionCoordinatorPlanPositionRepository.delete(oldInstitutionCoordinatorPlanPosition);
-                institutionPlanPosition.setAmountRequestedNet(institutionPlanPosition.getInstitutionCoordinatorPlanPositions().stream().map(InstitutionCoordinatorPlanPosition::getAmountRequestedNet).reduce(BigDecimal.ZERO, BigDecimal::add));
-                institutionPlanPosition.setAmountRequestedGross(institutionPlanPosition.getInstitutionCoordinatorPlanPositions().stream().map(InstitutionCoordinatorPlanPosition::getAmountRequestedGross).reduce(BigDecimal.ZERO, BigDecimal::add));
             } else {
                 Set<InstitutionCoordinatorPlanPosition> institutionCoordinatorPlanPositions = new HashSet<>();
                 InstitutionPublicProcurementPlanPosition institutionPlanPosition = new InstitutionPublicProcurementPlanPosition(
